requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pandas>=2.1.0
textblob>=0.17.1
nltk>=3.8.1
apscheduler>=3.10.4
colorlog>=6.8.0
pytest>=7.4.0
typing-extensions>=4.8.0
